Double type to Decimal

+++ OnlineCashRmk/Models/WriteofGood.cs
@@ -15,7 +15,7 @@ namespace OnlineCashRmk.Models
         public Good Good { get; set; }
         public string GoodName { get => Good?.Name; }
         public string GoodUnit { get => Good?.UnitDescription; }
-        public double Count { get; set; }
+        public decimal Count { get; set; }
         public decimal Price { get; set; }
         public decimal Sum { get => (decimal)Count * Price; }
     }
